chore: remove old types and implementations
