+++ src/providers/ergodex.ts
@@ -1,7 +1,9 @@
-import { Provider } from '@/types/core';
-import { timedCache } from '@/lib/utils';
 import { makeNativePools, makeTokenPools } from '@patternglobal/ergo-dex-sdk';
 import { AssetAmount, Explorer, RustModule } from '@patternglobal/ergo-sdk';
+
+import { asset2usd, timedCache } from '@/lib/utils';
+import { getNodeById } from '@/repositories/node';
+import { ArbitNodeId, Provider } from '@/types/core';
 
 await RustModule.load(true);
 
@@ -18,6 +20,25 @@ const getPool = timedCache(async (marketId: string) => {
 });
 
 const ErgoDex: Provider = {
+  id: 'ergodex',
+  name: 'ErgoDex',
+  type: 'real',
+  url: 'https://ergodex.io',
+
+  getExplicitFee: async (nodeId: ArbitNodeId, amounts: number[]) => {
+    const node = getNodeById(nodeId);
+    const erg = getNodeById('ergo:ERG');
+    const assetPrices = await asset2usd(node, amounts);
+    const [ergPrice] = await asset2usd(erg, [1_000000000]);
+    return assetPrices.map((amount) => {
+      const serviceFee = amount * 0.003 * ergPrice; // 0.3%
+      const maxExecutionFee = 0.0072 * ergPrice; // For Nitro of 1.2
+      const networkFee = 0.002 * ergPrice;
+
+      return serviceFee + maxExecutionFee + networkFee;
+    });
+  },
+
   async x2y(marketId: string, amounts: number[]) {
     const pool = await getPool(marketId);
     return amounts.map((amount) => {
