+++ src/repositories/edge.ts
@@ -1,5 +1,5 @@
 import edges from '@/data/edges';
-import { ArbitEdgeId } from '@/types/core-v2';
+import { ArbitEdgeId } from '@/types/core';
 
 export const getAllEdgeIds = () => {
   return Object.keys(edges);
