--- src/services/simple-arbitrategy.ts
@@ -1,5 +1,0 @@
-import simpleArbitrategy from '@/arbitrategy';
-import { computeArbitrategyProfit } from '@/lib/arbit-core';
-
-export const computeSimpleArbitrategyProfit = () =>
-  computeArbitrategyProfit(simpleArbitrategy);
