+++ src/utils/graph/getAdjacentNode.ts
@@ -1,4 +1,4 @@
-import { GraphEdge } from '@/types/core-v2';
+import { GraphEdge } from '@/types/core';
 
 export const getAdjacentNode = (node: string, edge: GraphEdge) => {
   return edge.nodes.x === node ? edge.nodes.y : edge.nodes.x;
